// src/Layout/Enums.ts
export enum LinearDirection {
  Horizontal = 'Horizontal',
  Vertical = 'Vertical',
}

export enum LinearJustify {
  Center = 'Center',
  Start = 'Start',
  End = 'End',
  SpaceBetween = 'SpaceBetween',
  SpaceAround = 'SpaceAround',
}

export enum LinearAlign {
  Center = 'Center',
  Start = 'Start',
  End = 'End',
  Stretch = 'Stretch',
}

export enum OverlayAlign {
  TopLeft = 'TopLeft',
  TopRight = 'TopRight',
  BottomLeft = 'BottomLeft',
  BottomRight = 'BottomRight',
}

export enum StackAlign {
  Center = 'Center',
  Start = 'Start',
  End = 'End',
}